Atualizado os pacotes do vue e colocado a api para funcionar com o docker compose

+++ RestWithASPNETudemy/Business/IBookBusiness.cs
@@ -7,11 +7,11 @@ namespace RestWithASPNETUdemy.Business
     public interface IBookBusiness
     {
         BookVO Create(BookVO book);
-        BookVO FindById(long id);
+        BookVO FindById(int id);
         IEnumerable<BookVO> FindAll();
         BookVO Update(BookVO book);
-        void Delete(long id);
-        bool Exists(long id);
+        void Delete(int id);
+        bool Exists(int id);
 
         PagedSearchVO<BookVO> FindWithPagedSearch(string title, string sortDirection, int pagesize, int page);
         IEnumerable<BookVO> FindByAuthor(string author);
