+++ RestWithASPNETudemy/Business/IPersonBusiness.cs
@@ -7,12 +7,12 @@ namespace RestWithASPNETUdemy.Business
     public interface IPersonBusiness
     {
          PersonVO Create(PersonVO person);
-         PersonVO FindById(long id);
+         PersonVO FindById(int id);
          IEnumerable<PersonVO> FindAll();
          PersonVO Update(PersonVO person);
-         void Delete(long id);
-         bool Exists(long id);
-         PersonVO Disable(long id);
+         void Delete(int id);
+         bool Exists(int id);
+         PersonVO Disable(int id);
          IEnumerable<PersonVO> FindByName(string name);
 
          PagedSearchVO<PersonVO> FindWithPagedSearch(string name, string direction, int pagesize, int page);
