+++ RestWithASPNETudemy/Repository/Generic/IBaseRepository.cs
@@ -8,12 +8,12 @@ namespace RestWithASPNETUdemy.Repository.Generic
     public interface IBaseRepository<T> where T: BaseEntity
     {
          T Create(T entity);
-         T FindById(long id);
+         T FindById(int id);
          List<T> FindAll();
          IEnumerable<T> Find(Expression<Func<T, bool>> filter = null);
          T Update(T entity);
-         void Delete(long id);
-         bool Exists(long? id);
+         void Delete(int id);
+         bool Exists(int? id);
 
          // Paginação
          List<T> FindWithPagedSearch(string query);
